And training forward slash to example URI for GitLab (#1507)
The example URI for creating an application in GitLab does not contain a trailing `/` but the Netlify CMS admin oauth implicit grant request to GitLab does include a trailing `/`.  This results in a failure with GitLab not matching the URI and returning an error. 

Following the example explicitly as most will do with copy-paste,  `https://example.com/app/admin` will be entered. Then the Netlify admin login will send `https://example.com/app/admin` and the URI will not match and GitLab will show an error indicating the URI is invalid.

The easy fix for the user is just to put the trailing `/` ... `https://example.com/app/admin/`.

I updated that example string in the hopes that users will not have to figure this out on their own in the future.